requirements.txt
pypresence>=4.6.0
pystray>=0.19.0
pillow>=10.0.0
yandex-music>=2.0.0
winrt-windows.media.control>=2.0.0
winrt-windows.storage.streams>=2.0.0
winrt-windows.foundation>=2.0.0
winrt-windows.foundation.collections>=2.0.0
pyinstaller>=6.0.0
